requirements.txt
openpyxl
markdown
pyyaml
customtkinter
pywin32
tkinterdnd2
